+++ Models/AvgBacklinksInfo.cs
@@ -18,48 +18,48 @@ namespace DataForSeo.Client.Models
         /// average number of backlinks
         /// </summary>
         [JsonProperty("backlinks", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public float? Backlinks { get; set; }
+        public double? Backlinks { get; set; }
 
         /// <summary>
         /// average number of dofollow links
         /// </summary>
         [JsonProperty("dofollow", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public float? Dofollow { get; set; }
+        public double? Dofollow { get; set; }
 
         /// <summary>
         /// average number of referring pages
         /// </summary>
         [JsonProperty("referring_pages", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public float? ReferringPages { get; set; }
+        public double? ReferringPages { get; set; }
 
         /// <summary>
         /// average number of referring domains
         /// </summary>
         [JsonProperty("referring_domains", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public float? ReferringDomains { get; set; }
+        public double? ReferringDomains { get; set; }
 
         /// <summary>
         /// average number of referring main domains
         /// </summary>
         [JsonProperty("referring_main_domains", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public float? ReferringMainDomains { get; set; }
+        public double? ReferringMainDomains { get; set; }
 
         /// <summary>
         /// average rank
         /// <br/>learn more about the metric and its calculation formula in this help center article
         /// </summary>
         [JsonProperty("rank", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public float? Rank { get; set; }
+        public double? Rank { get; set; }
 
         /// <summary>
         /// average main domain rank
         /// <br/>learn more about the metric and its calculation formula in this help center article
         /// </summary>
         [JsonProperty("main_domain_rank", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public float? MainDomainRank { get; set; }
+        public double? MainDomainRank { get; set; }
 
         /// <summary>
-        /// date and time when SERP data was updated
+        /// date and time when the dataset was updated
         /// <br/>in the UTC format: “yyyy-mm-dd hh-mm-ss +00:00”
         /// <br/>example:
         /// <br/>2019-11-15 12:57:46 +00:00
