+++ Models/RatingInfo.cs
@@ -31,7 +31,7 @@ namespace DataForSeo.Client.Models
         /// the maximum value for a rating_type
         /// </summary>
         [JsonProperty("rating_max", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public int? RatingMax { get; set; }
+        public double? RatingMax { get; set; }
 
         private IDictionary<string, object> _additionalProperties;
